remove TODO want "source identified" for var ptr *Source
var ptr *Source get optimized out during the SSA build
(see ssa/lift.go, line 131 for a starting point).

This seems fine to me as a nil source is not a source we need to
care about anyway. I added `var constPtr *Source` as an example to
explicitly show that we do not want a result there.

If the pointer is assigned a value though, then we do want a
report. To verify this case I have added an assignment to ptr.